+++ plone/app/discussion/browser/validator.py
@@ -11,7 +11,7 @@ from z3c.form.interfaces import IValidator
 from zope.component import adapts
 from zope.component import getMultiAdapter
 from zope.component import queryUtility
-from zope.interface import implements
+from zope.interface import implementer
 from zope.interface import Interface
 from zope.schema.interfaces import IField
 
@@ -32,8 +32,8 @@ except ImportError:
     pass
 
 
+@implementer(IValidator)
 class CaptchaValidator(validator.SimpleFieldValidator):
-    implements(IValidator)
     adapts(Interface, IDiscussionLayer, Interface, IField, Interface)
     #       Object, Request, Form, Field, Widget,
     # We adapt the CaptchaValidator class to all form fields (IField)